Unify Public & Private metadata into a single concern.
