+++ app/views/spree/admin/metadata/_action_buttons.html.erb
@@ -1,0 +1,3 @@
+<%= link_to Spree.t('admin.metadata.add_new_row'),
+      spree.create_metadata_admin_product_path(resource, kind),
+      class: 'btn btn-outline-secondary btn-sm', data: { turbo_method: :post } %>
