+++ app/views/spree/admin/metadata/_public_and_private.html.erb
@@ -1,0 +1,2 @@
+<%= render partial: 'spree/admin/metadata/form', locals: { resource: resource, f: f, kind: 'public' } %>
+<%= render partial: 'spree/admin/metadata/form', locals: { resource: resource, f: f, kind: 'private' } %>
