--- app/views/spree/admin/private_metadata/_action_buttons.html.erb
@@ -1,1 +1,0 @@
-<%= link_to Spree.t('admin.metadata.add_new_row'), spree.create_private_metadata_admin_product_path(resource), class: 'btn btn-outline-secondary btn-sm', data: { turbo_method: :post } %>
