--- app/views/spree/admin/private_metadata/_fields.html.erb
@@ -1,13 +1,0 @@
-<tr data-controller="required-input">
-  <td>
-    <%= text_field_tag("[private_metadata][key][]", nil, { id: nil, class: 'form-control', data: { required_input_target: "togglableInput" } }) %>
-  </td>
-  <td>
-    <%= text_field_tag("[private_metadata][value][]", nil, { id: nil, class: 'form-control', data: { action: "input->required-input#toggleRequired" } }) %>
-  </td>
-  <td class="text-center">
-    <%= link_to '#', class: 'btn btn-light btn-sm', data: { action: 'click->row#remove' }  do %>
-      <%= svg_icon name: "x-lg.svg", width: '16', height: '16', classes: 'p-0 mx-0' %>
-    <% end %>
-  </td>
-</tr>
