--- app/views/spree/admin/public_metadata/_form.html.erb
@@ -1,28 +1,0 @@
-<div class="card my-4">
-  <div class="card-header">
-    <div class="row pb-0">
-      <div class="col">
-        <h3 class="mb-0"><%= Spree.t('admin.metadata.public_metadata') %></h3>
-      </div>
-
-      <div class="col-auto text-right" id="add_public_metadata_fields_form">
-        <%= render partial: 'spree/admin/public_metadata/action_buttons', locals: { resource: resource } %>
-      </div>
-    </div>
-  </div>
-
-  <div class="card-body p-0">
-    <table class="table">
-      <thead>
-        <tr>
-          <th scope="col" ><%= raw(Spree.t('admin.metadata.key') + required_span_tag) %></th>
-          <th scope="col"><%= Spree.t('admin.metadata.value') %></th>
-          <th scope="col" style="width: 10%; text-align: center;"><%= Spree.t('admin.metadata.actions') %></th>
-        </tr>
-      </thead>
-        <%= render partial: 'spree/admin/public_metadata/table', locals: { resource: resource, f: f } %>
-    </table>
-  </div>
-</div>
-
-
