--- app/views/spree/admin/public_metadata/_table.html.erb
@@ -1,34 +1,0 @@
-<tbody id="public_metadata_form" data-controller="row">
-  <% if resource.public_metadata.any? %>
-    <% resource.public_metadata.each_with_index do |(key, value), index| %>
-      <tr id="public_metadata_row_<%= key %>">
-        <th scope="row">
-          <%= hidden_field_tag("[public_metadata][previous_key][]", key, { id: "public_metadata_previous_key_#{ index }" }) %>
-          <%= text_field_tag("[public_metadata][key][]", key, { id: "public_metadata_key_#{ index }",  class: 'form-control', required: true }) %>
-        </th>
-        <td>
-          <%= text_field_tag("[public_metadata][value][]", value, { id: "product_public_metadata_value_#{ index }",  class: 'form-control'}) %>
-        </td>
-        <td class="text-center">
-          <%= link_to spree.delete_public_metadata_admin_product_path(resource, key), class: 'btn button btn-danger btn-sm', data: { turbo_method: :delete, turbo_confirm: 'Are you sure?' } do %>
-            <%= svg_icon name: "delete.svg", width: '16', height: '16', classes: 'p-0 m-0 mb-1' %>
-          <% end %>
-        </td>
-      </tr>
-    <% end %>
-  <% else %>
-    <tr data-controller="required-input">
-      <td>
-        <%= text_field_tag("[public_metadata][key][]", nil, { id: nil, class: 'form-control', data: { required_input_target: "togglableInput" } }) %>
-      </td>
-      <td>
-        <%= text_field_tag("[public_metadata][value][]", nil, { id: nil, class: 'form-control', data: { action: "input->required-input#toggleRequired" } }) %>
-      </td>
-      <td class="text-center">
-        <%= link_to '#', class: 'btn btn-light btn-sm', data: { action: 'click->row#remove' }  do %>
-          <%= svg_icon name: "x-lg.svg", width: '16', height: '16', classes: 'p-0 mx-0' %>
-        <% end %>
-      </td>
-    </tr>
-  <% end %>
-</tbody>
